+++ R/airsis_parseData.R
@@ -68,20 +68,20 @@ airsis_parseData <- function(fileString) {
     
   } else if ( monitorType == "OTHER_1" ) {
     
-    logger.debug("Header line:\n\t%s", paste0(rawNames,collapse=','))
     logger.error("Older EBAM 1 file parsing is not supported")
+    logger.error("Header line:\n\t%s", paste0(rawNames,collapse=','))
     stop(paste0("Older EBAM 1 file parsing is not supported", call.=FALSE))
     
   } else if ( monitorType == "OTHER_2" ) {
     
-    logger.debug("Header line:\n\t%s", paste0(rawNames,collapse=','))
     logger.error("Older EBAM 2 file parsing is not supported")
+    logger.error("Header line:\n\t%s", paste0(rawNames,collapse=','))
     stop(paste0("Older EBAM 2 file parsing is not supported", call.=FALSE))
     
   } else {
     
-    logger.debug("Header line:\n\t%s", paste0(rawNames,collapse=','))
-    logger.error("Unkown file parsing is not supported")
+    logger.error("Unknown file parsing is not supported")
+    logger.error("Header line:\n\t%s", paste0(rawNames,collapse=','))
     stop(paste0("Unknown file parsing is not supported", call.=FALSE))
     
   }